install: fix software raid1 on esp
While running grub-install on an efi system where efi system partition
is configured as mdadm software raid1, it fails with errors like this:

  grub2-install: info: copying `/boot/grub2/x86_64-efi/core.efi' -> `/boot/efi/EFI/opensuse/grubx64.efi'.
  grub2-install: info: Registering with EFI: distributor = `opensuse', path = `\EFI\opensuse\grubx64.efi', ESP at mduuid/9182c46b9d469f79b48850b68f3371a5.
  grub2-install: info: executing efibootmgr --version </dev/null >/dev/null.
  grub2-install: info: executing modprobe -q efivars.
  grub2-install: info: executing efibootmgr -c -d.
  efibootmgr: option requires an argument -- 'd'
  efibootmgr version 14
  usage: efibootmgr [options]

This should work with mdadm raid1 with metadata 0.9 and 1.0 whose
superblocks are at the end of device. However
grub_install_register_efi() doesn't seem to work if the target is
multiple devices so that it errors out.

The patch changes grub_install_register_efi() to accept multiple devices
that can be used to creating efi boot entries for probed raid1 member
devices on mounted efi system partition.

This patch also adds check for metadata 0.9 or 1.0 or the validation
will fail to continue the install.

Signed-off-by: Michael Chang <mchang@suse.com>