diskfilter: use nodes in logical volume's segment as member device
Currently the grub_diskfilter_memberlist function returns all physical
volumes added to a volume group to which a logical volume (LV) belongs.
However this is suboptimal as it doesn't fit the intended behavior of
returning underlying devices that make up the LV. To give a clear
picture, the result should be identical to running commands below to
display the logical volumes with underlying physical volumes in use.

  lvs -o +devices /dev/system/root
  lvdisplay --maps /dev/system/root

This change is required if any part of the PV not used by root LV is
encrypted. Using this lvm setup as an example:

  localhost:~ # lsblk
  NAME            MAJ:MIN RM  SIZE RO TYPE  MOUNTPOINT
  vda             253:0    0   20G  0 disk
  ├─vda1          253:1    0    8M  0 part
  └─vda2          253:2    0   20G  0 part
    ├─system-swap 254:0    0  1.4G  0 lvm   [SWAP]
    └─system-root 254:1    0 18.6G  0 lvm   /
  vdb             253:16   0   10M  0 disk
  └─data          254:2    0    8M  0 crypt
    └─system-data 254:3    0    4M  0 lvm   /data

Running grub-install would end up with error because "system" VG
contains /dev/vdb which is encrypted.

  error: attempt to install to encrypted disk without cryptodisk
  enabled. Set `GRUB_ENABLE_CRYPTODISK=y' in file `/etc/default/grub'.

Certainly we can enable GRUB_ENABLE_CRYPTODISK=y and move on, but that
is not always acceptable since the server may need to boot unattended,
in addition typing passphase for every system startup can be a big
hassle of which most users would like to avoid.

This patch solves the problem by returning physical volumes, /dev/vda2,
rightly used by system-root in the example above, thus changing how
grub-install perceives the underlying block device to boot and avoids
the error from happening.

Signed-off-by: Michael Chang <mchang@suse.de>
Tested-by: Olav Reinert <seroton10@gmail.com>