refactor grub_install_register_efi to accept multiple disks
The grub_install_register_efi function is refactored to accept creating
efi boot variables for the device with more than one disk. It will be
labeled by distributor plus the name of the disk device in parentheses.

If the device has only one disk it will be label by distributor as
usual.

Signed-off-by: Michael Chang <mchang@suse.com>