Add boot entry for esp on raid 1
If the esp device is detected as raid 1, we test the driver name is
mdraid and it's metadata is formatted at the end of device by checking
the version to match 0.9 or 1.0.

Only if these conditions are met the raid 1 can be bootable in uefi
without extra support to translate the block mapping since it is 1:1
mapped. From that point on we can create boot variables to each
redundant esp from which grub is located and can be booted.

Signed-off-by: Michael Chang <mchang@suse.com>