linux: fix efi_relocate_kernel failure
With the dynamic allocated heap in new memory manager, it could use up all
usable memory with no reservation on subsequent kernel loading leading to
following error:

  EFI stub: ERROR: Failed to allocate usable memory for kernel
  EFI stub: ERROR: efi_relocate_kernel() failed!
  EFI stub: ERROR: efi_main failed!

The patch tries to returning the memory pages allocated by grub to the firmware
before handing over to linux kernel. This could eliminate the worry that we
have no limited amount of memory exclusively set for grub to guarentee the
memory requirement for booting subsequent component can be met.

Signed-off-by: Michael Chang <mchang@suse.com>