util/grub-wrap: New tool to wrap a file as a PE image
This new tool can wrap a non-PE file as a PE image, which can be
subsequently digitally signed (and verified by shim).

It's important that this type of PE images must be rejected when trying
to execute them. This is done by making them look like plain Windows
DLLs instead of UEFI images. Major differences between these wrappers
and normal UEFI images are:

 * DLL flag is set (nobody cares)
 * Entry point is set to 0 (shim rejects this)
 * Subsystem is set to Windows GUI instead of EFI Application (tianocore
   rejects this)

Signed-off-by: Zhang Boyang <zhangboyang.id@gmail.com>