tpm2: support unsealing key with authorized policy
To solve the PCR brittleness, TPM 2.0 allows the administrator to
'authorize' a new PCR policy, i.e. a new set of PCR values, to unseal
the existing key. This commit extends the SRK mode to support the
authorized policy mode to unseal the disk encryption key.

The usage of the authorized policy mode is very similar to the SRK mode
except two additional arguments: "-P" for the publicy key and "-S" for
the signed policy.

Example of the authorized policy mode:

tpm2_key_protector_init -m authpol -b sha256 -p 0,2,4,7 \
  -k (hd0,gpt1)/boot/grub2/sealed.key \
  -P (hd0,gpt1)/boot/grub2/pub.key \
  -S (hd0,gpt1)/boot/grub2/pol.sig

Signed-off-by: Gary Lin <glin@suse.com>